requirements.txt
Django==5.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
cryptography==41.0.7
